+++ apps/journeys-admin/src/components/AccessDialog/AccessDialog.tsx
@@ -1,29 +1,15 @@
-import { ReactElement, MouseEvent, useEffect, useState } from 'react'
-import Button from '@mui/material/Button'
-import Avatar from '@mui/material/Avatar'
-import Divider from '@mui/material/Divider'
-import Menu from '@mui/material/Menu'
 import List from '@mui/material/List'
-import MuiListItem from '@mui/material/ListItem'
-import ListItemAvatar from '@mui/material/ListItemAvatar'
-import ListItemText from '@mui/material/ListItemText'
-import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown'
+import { ReactElement, useEffect } from 'react'
+import ListItem from '@mui/material/ListItem'
 import { gql, useLazyQuery, useQuery } from '@apollo/client'
-import { compact } from 'lodash'
-import Skeleton from '@mui/material/Skeleton'
 import useMediaQuery from '@mui/material/useMediaQuery'
 import { Theme } from '@mui/material/styles'
 import { CopyTextField } from '@core/shared/ui/CopyTextField'
 import { Dialog } from '@core/shared/ui/Dialog'
-import {
-  GetJourneyWithUserJourneys,
-  GetJourneyWithUserJourneys_journey_userJourneys as UserJourney
-} from '../../../__generated__/GetJourneyWithUserJourneys'
+import { GetJourneyWithUserJourneys } from '../../../__generated__/GetJourneyWithUserJourneys'
 import { UserJourneyRole } from '../../../__generated__/globalTypes'
 import { GetCurrentUser } from '../../../__generated__/GetCurrentUser'
-import { RemoveUser } from './RemoveUser'
-import { ApproveUser } from './ApproveUser'
-import { PromoteUser } from './PromoteUser'
+import { UserJourneyList } from './UserJourneyList'
 
 export const GET_JOURNEY_WITH_USER_JOURNEYS = gql`
   query GetJourneyWithUserJourneys($id: ID!) {
@@ -95,7 +81,7 @@ export function AccessDialog({
       fullscreen={!smUp}
     >
       <List sx={{ pt: 0 }}>
-        <MuiListItem sx={{ p: 0 }}>
+        <ListItem sx={{ p: 0 }}>
           <CopyTextField
             value={
               typeof window !== 'undefined'
@@ -110,7 +96,7 @@ export function AccessDialog({
             helperText="Anyone with this link can see journey and ask for editing rights.
               You can accept or reject every request."
           />
-        </MuiListItem>
+        </ListItem>
         {!loading && (
           <UserJourneyList
             title="Requested Editing Rights"
@@ -130,135 +116,5 @@ export function AccessDialog({
         />
       </List>
     </Dialog>
-  )
-}
-
-interface UserJourneyListProps {
-  title: string
-  loading?: boolean
-  userJourneys?: UserJourney[] | null
-  disable: boolean
-}
-
-function UserJourneyList({
-  title,
-  loading,
-  userJourneys,
-  disable
-}: UserJourneyListProps): ReactElement {
-  return (
-    <>
-      {((userJourneys?.length != null && userJourneys.length > 0) ||
-        loading === true) && (
-        <>
-          <Divider sx={{ my: 2 }} />
-          <MuiListItem sx={{ px: 0 }}>{title}</MuiListItem>
-        </>
-      )}
-      {loading === true &&
-        [0, 1, 2].map((i) => (
-          <MuiListItem sx={{ px: 0 }} key={i}>
-            <ListItemAvatar>
-              <Skeleton variant="circular" width={40} height={40} />
-            </ListItemAvatar>
-            <ListItemText
-              primary={<Skeleton variant="text" width="60%" />}
-              secondary={<Skeleton variant="text" width="30%" />}
-            />
-          </MuiListItem>
-        ))}
-      {userJourneys?.map((userJourney) => (
-        <ListItem
-          key={userJourney.id}
-          userJourney={userJourney}
-          disabled={disable}
-        />
-      ))}
-    </>
-  )
-}
-
-interface ListItemProps {
-  userJourney: UserJourney
-  disabled: boolean
-}
-
-function ListItem({ userJourney, disabled }: ListItemProps): ReactElement {
-  const [anchorEl, setAnchorEl] = useState<HTMLElement | null>(null)
-  const open = Boolean(anchorEl)
-  const { id, role, user } = userJourney
-
-  useEffect(() => {
-    return () => {
-      setAnchorEl(null)
-    }
-  }, [])
-
-  const handleClick = (event: MouseEvent<HTMLElement>): void => {
-    setAnchorEl(event.currentTarget)
-  }
-  const handleClose = (): void => {
-    setAnchorEl(null)
-  }
-
-  const displayName = compact([user?.firstName, user?.lastName]).join(' ')
-
-  return (
-    <>
-      <MuiListItem
-        sx={{
-          px: 0,
-          '& > .MuiListItemSecondaryAction-root': {
-            right: 0
-          }
-        }}
-        secondaryAction={
-          <Button
-            aria-controls={open ? 'basic-menu' : undefined}
-            aria-haspopup="true"
-            aria-expanded={open ? 'true' : undefined}
-            onClick={handleClick}
-            disabled={disabled || role === 'owner'}
-            endIcon={<ArrowDropDownIcon />}
-            sx={{
-              color: 'text.primary',
-              typography: 'body2'
-            }}
-          >
-            {role === 'inviteRequested' && 'Manage'}
-            {role === 'owner' && 'Owner'}
-            {role === 'editor' && 'Editor'}
-          </Button>
-        }
-      >
-        <ListItemAvatar>
-          <Avatar src={user?.imageUrl ?? undefined} alt={displayName}>
-            {displayName.charAt(0)?.toUpperCase()}
-          </Avatar>
-        </ListItemAvatar>
-        <ListItemText primary={displayName} secondary={user?.email} />
-      </MuiListItem>
-      <Menu
-        anchorEl={anchorEl}
-        open={open}
-        onClose={handleClose}
-        anchorOrigin={{
-          vertical: 'bottom',
-          horizontal: 'right'
-        }}
-        transformOrigin={{
-          vertical: 'top',
-          horizontal: 'right'
-        }}
-      >
-        {role === 'inviteRequested' ? (
-          <ApproveUser id={id} onClick={handleClose} />
-        ) : (
-          <PromoteUser id={id} onClick={handleClose} />
-        )}
-        <Divider />
-        <RemoveUser id={id} onClick={handleClose} />
-      </Menu>
-    </>
   )
 }
