+++ apps/journeys-admin/src/components/AccessDialog/UserJourneyList/UserJourneyItem/index.ts
@@ -1,0 +1,1 @@
+export { UserJourneyItem } from './UserJourneyItem'
