+++ apps/journeys-admin/src/components/AccessDialog/UserJourneyList/index.ts
@@ -1,0 +1,1 @@
+export { UserJourneyList } from './UserJourneyList'
